[DYNAREC] Fixed some internal structure size

+++ src/dynarec/dynarec_arm_private.h
@@ -18,12 +18,12 @@ typedef struct dynablock_s dynablock_t;
 typedef union neon_cache_s {
     int8_t           v;
     struct {
-        unsigned int t:4;   // reg type
-        unsigned int n:4;   // reg number
+        uint8_t t:4;   // reg type
+        uint8_t n:4;   // reg number
     };
 } neon_cache_t;
 typedef union sse_cache_s {
-    int     v;
+    int8_t     v;
     struct {
         uint8_t reg;
         uint8_t write;
